API: Extract methods from handler into reusable location
Pull all the methods into an interface/callable methods.
This allows us to better support a command console like
other cryptos.  This could also be used later with the CLI.

Also fix typo in Transaction builder.
