+++ src/main/java/org/semux/api/ApiHandlerImpl.java
@@ -6,60 +6,22 @@
  */
 package org.semux.api;
 
-import java.net.UnknownHostException;
-import java.util.Arrays;
 import java.util.Map;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
-import java.util.stream.Collectors;
-
-import org.semux.Kernel;
-import org.semux.api.response.AddNodeResponse;
-import org.semux.api.response.CreateAccountResponse;
-import org.semux.api.response.DoTransactionResponse;
-import org.semux.api.response.GetAccountResponse;
-import org.semux.api.response.GetAccountTransactionsResponse;
-import org.semux.api.response.GetBlockResponse;
-import org.semux.api.response.GetDelegateResponse;
-import org.semux.api.response.GetDelegatesResponse;
-import org.semux.api.response.GetInfoResponse;
-import org.semux.api.response.GetLatestBlockNumberResponse;
-import org.semux.api.response.GetLatestBlockResponse;
-import org.semux.api.response.GetPeersResponse;
-import org.semux.api.response.GetPendingTransactionsResponse;
-import org.semux.api.response.GetRootResponse;
-import org.semux.api.response.GetTransactionLimitsResponse;
-import org.semux.api.response.GetTransactionResponse;
-import org.semux.api.response.GetValidatorsResponse;
-import org.semux.api.response.GetVoteResponse;
-import org.semux.api.response.GetVotesResponse;
-import org.semux.api.response.ListAccountsResponse;
-import org.semux.api.response.SendTransactionResponse;
-import org.semux.api.util.TransactionBuilder;
-import org.semux.core.Block;
-import org.semux.core.BlockchainImpl;
-import org.semux.core.PendingManager;
-import org.semux.core.Transaction;
-import org.semux.core.TransactionType;
-import org.semux.core.exception.WalletLockedException;
-import org.semux.core.state.Account;
-import org.semux.core.state.Delegate;
-import org.semux.crypto.CryptoException;
-import org.semux.crypto.EdDSA;
-import org.semux.crypto.Hex;
-import org.semux.net.NodeManager.Node;
-import org.semux.util.exception.UnreachableException;
 
 import io.netty.handler.codec.http.HttpHeaders;
+import org.semux.Kernel;
+import org.semux.api.response.GetRootResponse;
+import org.semux.util.exception.UnreachableException;
 
 /**
  * Semux RESTful API handler implementation.
- * 
+ *
  * TODO: Auto-generate API docs
  */
 public class ApiHandlerImpl implements ApiHandler {
 
     private final Kernel kernel;
+    private final SemuxApi semuxApi;
 
     /**
      * Creates an API handler.
@@ -68,6 +30,7 @@ public class ApiHandlerImpl implements ApiHandler {
      */
     public ApiHandlerImpl(Kernel kernel) {
         this.kernel = kernel;
+        this.semuxApi = new SemuxApiImpl(kernel);
     }
 
     @Override
@@ -166,7 +129,7 @@ public class ApiHandlerImpl implements ApiHandler {
      * @return
      */
     private ApiHandlerResponse getInfo() {
-        return new GetInfoResponse(true, new GetInfoResponse.Result(kernel));
+        return semuxApi.getInfo();
     }
 
     /**
@@ -175,10 +138,7 @@ public class ApiHandlerImpl implements ApiHandler {
      * @return
      */
     private ApiHandlerResponse getPeers() {
-        return new GetPeersResponse(true,
-                kernel.getChannelManager().getActivePeers().parallelStream()
-                        .map(GetPeersResponse.Result::new)
-                        .collect(Collectors.toList()));
+       return semuxApi.getPeers();
     }
 
     /**
@@ -188,32 +148,7 @@ public class ApiHandlerImpl implements ApiHandler {
      * @return result
      */
     private ApiHandlerResponse addNode(Map<String, String> params) {
-        try {
-            kernel.getNodeManager().addNode(validateAddNodeParameter(params.get("node")));
-            return new AddNodeResponse(true);
-        } catch (IllegalArgumentException e) {
-            return failure(e.getMessage());
-        }
-    }
-
-    /**
-     * Validate node parameter of /add_node API
-     *
-     * @param node
-     *            node parameter of /add_node API
-     * @return validated hostname and port number
-     */
-    private Node validateAddNodeParameter(String node) {
-        if (node == null || node.length() == 0) {
-            throw new IllegalArgumentException("Parameter `node` can't be empty");
-        }
-
-        Matcher matcher = Pattern.compile("^(?<host>.+?):(?<port>\\d+)$").matcher(node.trim());
-        if (!matcher.matches()) {
-            throw new IllegalArgumentException("Parameter `node` must in format of `host:port`");
-        }
-
-        return new Node(matcher.group("host"), Integer.parseInt(matcher.group("port")));
+       return semuxApi.addNode(params.get("node"));
     }
 
     /**
@@ -225,21 +160,14 @@ public class ApiHandlerImpl implements ApiHandler {
     private ApiHandlerResponse getBlock(Map<String, String> params) {
         String number = params.get("number");
         String hash = params.get("hash");
-        Block block;
 
         if (number != null) {
-            block = kernel.getBlockchain().getBlock(Long.parseLong(number));
+            return semuxApi.getBlock(Long.parseLong(number));
         } else if (hash != null) {
-            block = kernel.getBlockchain().getBlock(Hex.decode0x(hash));
+            return semuxApi.getBlock(hash);
         } else {
-            return failure("Either parameter `number` or `hash` has to be provided");
+            return semuxApi.failure("Either parameter `number` or `hash` has to be provided");
         }
-
-        if (block == null) {
-            return failure("The requested block was not found");
-        }
-
-        return new GetBlockResponse(true, new GetBlockResponse.Result(block));
     }
 
     /**
@@ -248,11 +176,7 @@ public class ApiHandlerImpl implements ApiHandler {
      * @return
      */
     private ApiHandlerResponse getPendingTransactions() {
-        return new GetPendingTransactionsResponse(true,
-                kernel.getPendingManager().getTransactions().parallelStream()
-                        .map(pendingTransaction -> pendingTransaction.transaction)
-                        .map(GetTransactionResponse.Result::new)
-                        .collect(Collectors.toList()));
+        return semuxApi.getPendingTransactions();
     }
 
     /**
@@ -265,36 +189,8 @@ public class ApiHandlerImpl implements ApiHandler {
         String address = params.get("address");
         String from = params.get("from");
         String to = params.get("to");
-        byte[] addressBytes;
-        int fromInt;
-        int toInt;
 
-        if (address == null) {
-            return failure("Parameter `address` is required");
-        }
-
-        try {
-            addressBytes = Hex.decode0x(address);
-        } catch (CryptoException ex) {
-            return failure("Parameter `address` is not a valid hexadecimal string");
-        }
-
-        try {
-            fromInt = Integer.parseInt(from);
-        } catch (NumberFormatException ex) {
-            return failure("Parameter `from` is not a valid integer");
-        }
-
-        try {
-            toInt = Integer.parseInt(to);
-        } catch (NumberFormatException ex) {
-            return failure("Parameter `to` is not a valid integer");
-        }
-
-        return new GetAccountTransactionsResponse(true,
-                kernel.getBlockchain().getTransactions(addressBytes, fromInt, toInt).parallelStream()
-                        .map(GetTransactionResponse.Result::new)
-                        .collect(Collectors.toList()));
+        return semuxApi.getAccountTransactions(address, from, to);
     }
 
     /**
@@ -305,23 +201,7 @@ public class ApiHandlerImpl implements ApiHandler {
      */
     private ApiHandlerResponse getTransaction(Map<String, String> params) {
         String hash = params.get("hash");
-        if (hash == null) {
-            return failure("Parameter `hash` can't be null");
-        }
-
-        byte[] hashBytes;
-        try {
-            hashBytes = Hex.decode0x(hash);
-        } catch (CryptoException ex) {
-            return failure("Parameter `hash` is not a valid hexadecimal string");
-        }
-
-        Transaction transaction = kernel.getBlockchain().getTransaction(hashBytes);
-        if (transaction == null) {
-            return failure("The request transaction was not found");
-        }
-
-        return new GetTransactionResponse(true, new GetTransactionResponse.Result(transaction));
+        return semuxApi.getTransaction(hash);
     }
 
     /**
@@ -331,17 +211,8 @@ public class ApiHandlerImpl implements ApiHandler {
      * @return
      */
     private ApiHandlerResponse sendTransaction(Map<String, String> params) {
-        try {
-            String raw = params.get("raw");
-            if (raw == null) {
-                return failure("Parameter `raw` can't be null");
-            }
-
-            kernel.getPendingManager().addTransaction(Transaction.fromBytes(Hex.decode0x(raw)));
-            return new SendTransactionResponse(true);
-        } catch (CryptoException e) {
-            return failure("Parameter `raw` is not a valid hexadecimal string");
-        }
+        String raw = params.get("raw");
+        return semuxApi.sendTransaction(raw);
     }
 
     /**
@@ -352,19 +223,7 @@ public class ApiHandlerImpl implements ApiHandler {
      */
     private ApiHandlerResponse getAccount(Map<String, String> params) {
         String address = params.get("address");
-        if (address == null) {
-            return failure("Parameter `address` can't be null");
-        }
-
-        byte[] addressBytes;
-        try {
-            addressBytes = Hex.decode0x(address);
-        } catch (CryptoException ex) {
-            return failure("Parameter `address` is not a valid hexadecimal string");
-        }
-
-        Account account = kernel.getBlockchain().getAccountState().getAccount(addressBytes);
-        return new GetAccountResponse(true, new GetAccountResponse.Result(account));
+        return semuxApi.getAccount(address);
     }
 
     /**
@@ -374,18 +233,8 @@ public class ApiHandlerImpl implements ApiHandler {
      * @return
      */
     private ApiHandlerResponse addToBlackList(Map<String, String> params) {
-        try {
-            String ip = params.get("ip");
-            if (ip == null || ip.trim().length() == 0) {
-                return failure("Parameter `ip` can't be empty");
-            }
-
-            kernel.getChannelManager().getIpFilter().blacklistIp(ip.trim());
-
-            return new ApiHandlerResponse(true, null);
-        } catch (UnknownHostException | IllegalArgumentException ex) {
-            return failure(ex.getMessage());
-        }
+        String ip = params.get("ip");
+        return semuxApi.addToBlacklist(ip);
     }
 
     /**
@@ -395,18 +244,8 @@ public class ApiHandlerImpl implements ApiHandler {
      * @return
      */
     private ApiHandlerResponse addToWhiteList(Map<String, String> params) {
-        try {
-            String ip = params.get("ip");
-            if (ip == null || ip.trim().length() == 0) {
-                return failure("Parameter `ip` can't be empty");
-            }
-
-            kernel.getChannelManager().getIpFilter().whitelistIp(ip.trim());
-
-            return new ApiHandlerResponse(true, null);
-        } catch (UnknownHostException | IllegalArgumentException ex) {
-            return failure(ex.getMessage());
-        }
+        String ip = params.get("ip");
+        return semuxApi.addToWhitelist(ip);
     }
 
     /**
@@ -415,7 +254,7 @@ public class ApiHandlerImpl implements ApiHandler {
      * @return
      */
     private ApiHandlerResponse getLatestBlockNumber() {
-        return new GetLatestBlockNumberResponse(true, kernel.getBlockchain().getLatestBlockNumber());
+        return semuxApi.getLatestBlockNumber();
     }
 
     /**
@@ -424,7 +263,7 @@ public class ApiHandlerImpl implements ApiHandler {
      * @return
      */
     private ApiHandlerResponse getLatestBlock() {
-        return new GetLatestBlockResponse(true, new GetBlockResponse.Result(kernel.getBlockchain().getLatestBlock()));
+        return semuxApi.getLatestBlock();
     }
 
     /**
@@ -435,25 +274,7 @@ public class ApiHandlerImpl implements ApiHandler {
      */
     private ApiHandlerResponse getDelegate(Map<String, String> params) {
         String address = params.get("address");
-        if (address == null) {
-            return failure("Parameter `address` can't be null");
-        }
-
-        byte[] addressBytes;
-        try {
-            addressBytes = Hex.decode0x(address);
-        } catch (CryptoException e) {
-            return failure(e.getMessage());
-        }
-
-        Delegate delegate = kernel.getBlockchain().getDelegateState().getDelegateByAddress(addressBytes);
-        if (delegate == null) {
-            return failure("The provided address is not a delegate");
-        }
-
-        BlockchainImpl.ValidatorStats validatorStats = kernel.getBlockchain().getValidatorStats(addressBytes);
-
-        return new GetDelegateResponse(true, new GetDelegateResponse.Result(validatorStats, delegate));
+        return semuxApi.getDelegate(address);
     }
 
     /**
@@ -462,8 +283,7 @@ public class ApiHandlerImpl implements ApiHandler {
      * @return
      */
     private ApiHandlerResponse getValidators() {
-        return new GetValidatorsResponse(true, kernel.getBlockchain().getValidators().parallelStream()
-                .map(v -> Hex.PREF + v).collect(Collectors.toList()));
+       return semuxApi.getValidators();
     }
 
     /**
@@ -472,11 +292,7 @@ public class ApiHandlerImpl implements ApiHandler {
      * @return
      */
     private ApiHandlerResponse getDelegates() {
-        return new GetDelegatesResponse(true,
-                kernel.getBlockchain().getDelegateState().getDelegates().parallelStream()
-                        .map(delegate -> new GetDelegateResponse.Result(
-                                kernel.getBlockchain().getValidatorStats(delegate.getAddress()), delegate))
-                        .collect(Collectors.toList()));
+        return semuxApi.getDelegates();
     }
 
     /**
@@ -488,30 +304,7 @@ public class ApiHandlerImpl implements ApiHandler {
     private ApiHandlerResponse getVote(Map<String, String> params) {
         String voter = params.get("voter");
         String delegate = params.get("delegate");
-        byte[] voterBytes;
-        byte[] delegateBytes;
-
-        if (voter == null) {
-            return failure("Parameter `voter` is required");
-        }
-
-        try {
-            voterBytes = Hex.decode0x(voter);
-        } catch (CryptoException ex) {
-            return failure("Parameter `voter` is not a valid hexadecimal string");
-        }
-
-        if (delegate == null) {
-            return failure("Parameter `delegate` is required");
-        }
-
-        try {
-            delegateBytes = Hex.decode0x(delegate);
-        } catch (CryptoException ex) {
-            return failure("Parameter `delegate` is not a valid hexadecimal string");
-        }
-
-        return new GetVoteResponse(true, kernel.getBlockchain().getDelegateState().getVote(voterBytes, delegateBytes));
+        return semuxApi.getVotes(delegate, voter);
     }
 
     /**
@@ -522,20 +315,7 @@ public class ApiHandlerImpl implements ApiHandler {
      */
     private ApiHandlerResponse getVotes(Map<String, String> params) {
         String delegate = params.get("delegate");
-        if (delegate == null) {
-            return failure("Parameter `delegate` can't be null");
-        }
-
-        byte[] delegateBytes;
-        try {
-            delegateBytes = Hex.decode0x(delegate);
-        } catch (CryptoException ex) {
-            return failure("Parameter `delegate` is not a valid hexadecimal string");
-        }
-
-        return new GetVotesResponse(true,
-                kernel.getBlockchain().getDelegateState().getVotes(delegateBytes).entrySet().parallelStream()
-                        .collect(Collectors.toMap(entry -> Hex.PREF + entry.getKey().toString(), Map.Entry::getValue)));
+        return semuxApi.getVotes(delegate);
     }
 
     /**
@@ -544,9 +324,7 @@ public class ApiHandlerImpl implements ApiHandler {
      * @return
      */
     private ApiHandlerResponse listAccounts() {
-        return new ListAccountsResponse(true, kernel.getWallet().getAccounts().parallelStream()
-                .map(acc -> Hex.PREF + acc.toAddressString())
-                .collect(Collectors.toList()));
+       return semuxApi.listAccounts();
     }
 
     /**
@@ -555,14 +333,7 @@ public class ApiHandlerImpl implements ApiHandler {
      * @return
      */
     private ApiHandlerResponse createAccount() {
-        try {
-            EdDSA key = new EdDSA();
-            kernel.getWallet().addAccount(key);
-            kernel.getWallet().flush();
-            return new CreateAccountResponse(true, Hex.PREF + key.toAddressString());
-        } catch (WalletLockedException e) {
-            return failure(e.getMessage());
-        }
+        return semuxApi.createAccount();
     }
 
     /**
@@ -572,15 +343,7 @@ public class ApiHandlerImpl implements ApiHandler {
      * @return
      */
     private ApiHandlerResponse getTransactionLimits(Map<String, String> params) {
-        try {
-            return new GetTransactionLimitsResponse(kernel, TransactionType.valueOf(params.get("type")));
-        } catch (NullPointerException | IllegalArgumentException e) {
-            return failure(String.format(
-                    "Invalid transaction type. (must be one of %s)",
-                    Arrays.stream(TransactionType.values())
-                            .map(TransactionType::toString)
-                            .collect(Collectors.joining(","))));
-        }
+        return semuxApi.getTransactionLimits(params.get("type"));
     }
 
     /**
@@ -604,44 +367,24 @@ public class ApiHandlerImpl implements ApiHandler {
             return failure("Wallet is locked");
         }
 
+        String from = params.get("from");
+        String to = params.get("to");
+        String value = params.get("value");
+        String fee = params.get("fee");
+        String data = params.get("data");
+
         // [2] parse transaction type
-        TransactionType type;
         switch (cmd) {
         case TRANSFER:
-            type = TransactionType.TRANSFER;
-            break;
+            return semuxApi.transfer(value,from,to,fee,data);
         case DELEGATE:
-            type = TransactionType.DELEGATE;
-            break;
+            return semuxApi.registerDelegate(from,fee,data);
         case VOTE:
-            type = TransactionType.VOTE;
-            break;
+            return semuxApi.vote(from,to,value,fee);
         case UNVOTE:
-            type = TransactionType.UNVOTE;
-            break;
+            return semuxApi.unvote(from,to,value,fee);
         default:
             return failure("Unsupported transaction type: " + cmd.toString());
-        }
-
-        // [3] build and send the transaction to PendingManager
-        try {
-            TransactionBuilder transactionBuilder = new TransactionBuilder(kernel, type)
-                    .withFrom(params.get("from"))
-                    .withTo(params.get("to"))
-                    .withValue(params.get("value"))
-                    .withFee(params.get("fee"))
-                    .withData(params.get("data"));
-
-            Transaction tx = transactionBuilder.build();
-
-            PendingManager.ProcessTransactionResult result = kernel.getPendingManager().addTransactionSync(tx);
-            if (result.error != null) {
-                return failure("Transaction rejected by pending manager: " + result.error.toString());
-            }
-
-            return new DoTransactionResponse(true, null, Hex.encode0x(tx.getHash()));
-        } catch (IllegalArgumentException ex) {
-            return failure(ex.getMessage());
         }
     }
